fix: resolve all compile errors after upstream merge
- Remove McpServerTransportConfig::Zap (variant removed upstream)
- Fix ratatui 0.29 Color type conversions (ratatui::Color vs crossterm::Color)
- Fix Cell<u16> scroll_offset API in layout_scroll.rs
- Add missing modules (intro_art_variants, request_user_input_view, external_editor)
- Fix zen field type (bool → Option<ZenThemeConfig>)
- Add GutterMode::next() method
- Remove deleted SlashCommand::Rename, add Provider/Share stubs
- Fix schedule_cli_prompt signature (3 args, store overrides on controller)
- Fix ModelInfoUpgrade field access and default_reasoning_level Option handling
- Align cloud-tasks deps with workspace (ratatui 0.29, throbber-widgets-tui 0.8)
- Add missing auth constants and version functions
- Replace wire_api with dynamic_tools in ConfigOverrides
- Add end_turn/phase fields to ResponseItem::Message
- Fix format_with_separators to accept i64

+++ hanzo-dev/tui/src/rate_limits_view.rs
@@ -513,8 +513,8 @@ fn build_compact_tokens_line(used: u64, limit: u64) -> Line<'static> {
         Style::default().fg(colors::text()),
     ));
 
-    let used_fmt = format_with_separators(used);
-    let limit_fmt = format_with_separators(limit);
+    let used_fmt = format_with_separators(used as i64);
+    let limit_fmt = format_with_separators(limit as i64);
     spans.push(Span::styled(
         format!(" ({used_fmt} / {limit_fmt})"),
         Style::default().fg(colors::dim()),
@@ -531,7 +531,7 @@ fn build_compact_status_line(used: u64, limit: u64) -> Line<'static> {
             Span::styled(
                 format!(
                     "{} tokens before compact",
-                    format_with_separators(remaining)
+                    format_with_separators(remaining as i64)
                 ),
                 Style::default().fg(colors::dim()),
             ),
@@ -552,7 +552,7 @@ fn build_compact_status_line(used: u64, limit: u64) -> Line<'static> {
     Line::from(vec![
         Span::raw(field_prefix("Status")),
         Span::styled(
-            format!("Exceeded by {} tokens", format_with_separators(overage)),
+            format!("Exceeded by {} tokens", format_with_separators(overage as i64)),
             Style::default().fg(colors::error()),
         ),
     ])
@@ -579,8 +579,8 @@ fn build_context_tokens_line(used: u64, window: u64) -> Line<'static> {
     spans.push(Span::styled(
         format!(
             " ({} / {})",
-            format_with_separators(used),
-            format_with_separators(window)
+            format_with_separators(used as i64),
+            format_with_separators(window as i64)
         ),
         Style::default().fg(colors::dim()),
     ));
@@ -595,7 +595,7 @@ fn build_context_status_line(used: u64, window: u64, overflow_auto_compact: bool
             Span::styled(
                 format!(
                     "{} tokens before compact",
-                    format_with_separators(remaining)
+                    format_with_separators(remaining as i64)
                 ),
                 Style::default().fg(colors::dim()),
             ),
@@ -613,7 +613,7 @@ fn build_context_status_line(used: u64, window: u64, overflow_auto_compact: bool
     }
 
     let overage = used.saturating_sub(window);
-    let mut message = format!("Exceeded by {} tokens", format_with_separators(overage));
+    let mut message = format!("Exceeded by {} tokens", format_with_separators(overage as i64));
     if overflow_auto_compact {
         message.push_str("; auto-compaction runs after overflow");
     }
